requirements.txt
pandas
numpy
streamlit
streamlit-authenticator
pyyaml
openpyxl